pyproject.toml
[tool.poetry]
name = "cloud-build"
description = ""
version = "0.1.0"
authors = ["sfujiwara <shuhei.fujiwara@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.7"
google-cloud-build = "2.0.0"
# google-cloud-build = "1.1.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.black]
line-length = 110
